D3Bindings Refactor
D3Bindings are no longer based on string literals. Instead they are
based on Enums. The Relevant forms class also uses this enum to
determine how many fields to create as well. This is way more scalable
and much more efficient. Instead of creating 12 fields manually, they
are generated based on the amount of enums based on a default label and
text field. However, this could be improved even more if I had a better
grasp on the code base. What actually needs to happen is the gamepad
needs to send out any combination of input to the computer not just a
single key as well as some sort of held functinality. The way bindings
are set up currently is just sloppy and I'll have to rework the whole
system as some point once I dive into the rest of the project.