HotKey
Fixed hotkey bindings to not use it's own class. I'm sure the class is
useful in some cases, but it's overly complex and didn't seem to work
right for me for whatever reason. It only takes about 10 lines of code
to recreate what it's doing inside of the HUDForm instead using the same
.dll so I don't see a reason to separate it out for now. Maybe later,
but even if I do it should absolutly be in a static class, not a new
object for each hotkey.