Process Inputs And Basic Mouse
Added implementation for processing inputs using XInput based on the
configuration set by the user. Added basic mouse functionality.

Added some basic indication to the user that a setting was changed and
needs to be saved by using an asterisk '*' next to the save button so
long as something hasn't been saved.