AutoHotkey
So there is a C# wrapper for using AutoHotkey. It's stupid powerful. You
can simply pass it a string and it executes it as if it was a script.
This means I can get the same behavior of AutoHotkey's button presses in
my C# form. Works great so far but need to test out performace to see if
there are any downsides. If not, this is amazing. It means I can make it
so that a user can create a small script in windows form and bind it to
a key if they want to instead of a standard binding. Pretty crazy if you
ask me. This would give a crazy amount of user control to the
application.

Implemented use of the AutoHotkey wrapper for the VirtualKeyboard class.
More work is needed for parsing the missing 'System.Windows.Forms.Keys'
into the AutoHotkey equivilant.