Finished BinidingType
You can now configure the stick values, trigger values, and the new
BindingType enum which shows different Controls depending on the enum.
Script is a multiline text box, Keys is the standard key parser text
box, and Special provides a combo box which has yet to be filled with
values and programmed.

Next up is adding the functionality to the input manager