Working Binding Control edit check
Have a working system for checking to see if a binding control has been edited and needs to be saved. Not sure how I'm going to have it pass through to the main config form, but this is a good first step. There's a lot of spagetti code it seems with making sure the original and temp never get improperly synced, but honestly I'm not sure of another way to do it other than an action queue which sounds completely not needed. I wouldn't even know where to start to be honest, but I feel like that's how some apps check for modifications. It's absolutely how they can do undo and redo. Maybe that's something I'd like to implement, but I really should get base stuff completely done first before something like an undo button--however useful it may be.